requirements.txt
bleach>=6.0,<7
flask>=2.3,<4
markdown-it-py>=3.0,<4
mdit-py-plugins>=0.4,<1
